refactor: share variant styles and relocate Callouts slot to layout/
