+++ resources/js/layout/components.ts
@@ -1,6 +1,6 @@
 import { createPlugin, eagerComponent } from "@lattice-php/lattice/core/registry";
 import BreadcrumbsComponent from "./breadcrumbs";
-import CalloutsComponent from "../toast/callouts";
+import CalloutsComponent from "./callouts";
 import DropdownComponent from "./dropdown";
 import MenuComponent from "./menu";
 import MenuItemComponent from "./menu-item";
