+++ resources/js/layout/index.ts
@@ -1,3 +1,4 @@
+export { default as Callouts } from "./callouts";
 export { layoutComponents } from "./components";
 export { OutletContext, SidebarCollapsedContext, useSidebarCollapsed } from "./context";
 export { default as SchemaLayout } from "./schema-layout";
