+++ resources/js/toast/index.ts
@@ -1,4 +1,3 @@
-export { default as Callouts } from "./callouts";
 export { normalizeCallout, onCallout } from "./callout";
 export type { Callout } from "./callout";
 export { Toaster } from "./toaster";
